Add 'checking'
Added GameStateManager.java which is responsible for checking if the players is in 'Check' (and eventually in check-mate). If player is in check after executing a move, the move will be undone and move counted as illegal.
Added GameplayManager.java which handles gameplay (manages executing moves)
Expanded on various functionalities as well as made the code more modular.